refactor(qwk): extract NDX record encoder into ndx.go
Move makeNDXRecord and its MBF4/MSBIN4 doc-comment block from writer.go
into a new focused file ndx.go. Removes the math import from writer.go.
writer.go is now 282 lines (under the 300-line ceiling).
